Convert run_worker_threads to a callable class
The run_worker_threads() functions had a C-style static variable to
ensure that a thread name would not be reused. In other words,
subsequent calls with the same name_prefix would result in threads with
unique names. See the below unit test taken from tests/test_threading.py
for an example:

def test_worker_thread_names():
    thread_names = set()

    def add_name_to_list():
        thread_names.add(current_thread().name)

    # the num_workers parameter defaults to 2
    run_worker_threads(target=add_name_to_list, name_prefix="A")
    run_worker_threads(target=add_name_to_list, name_prefix="B")
    run_worker_threads(target=add_name_to_list, name_prefix="A")

    assert "A-01" in thread_names
    assert "A-02" in thread_names
    assert "A-03" in thread_names
    assert "A-04" in thread_names
    assert "B-01" in thread_names
    assert "B-02" in thread_names
    assert len(thread_names) == 6

The C-style static variable tripped up mypy and is widely regarded as
"not pythonic", for whatever that's worth. Refactoring the function to
be a callable object resolves the objections of mypy and the
(theoretical) community.