Use a list instead of a set in test_worker_thread_names()
The use of a set count potentially mask an issue where duplicate thread
names were used, because a set will only allow one instance of a string,
multiple identical strings can be added but still only appear once in
the set. Changing from a set to a list fixes this.